requirements.txt
transformers
datasets
torch
Pillow